clean: Consolidate 'copyable' things into a `CopyButton` component
